Refactor main.rb by creating helper.rb
